Move some helper functions, add comments, update documentation
